+++ src/ManifoldFields.jl
@@ -6,6 +6,7 @@ using StaticArrays
 using YAXArrays
 
 export DiscreteField, mesh, data, location, withmesh
+export FieldSet, fields, field_names
 export location_dimname, expected_location_length
 export interpolate
 export UGridDataset, to_ugrid, from_ugrid, from_ugrid_mesh, save_ugrid, load_ugrid,
@@ -13,7 +14,9 @@ export UGridDataset, to_ugrid, from_ugrid, from_ugrid_mesh, save_ugrid, load_ugr
 
 include("locations.jl")
 include("field.jl")
+include("fieldset.jl")
 include("broadcast.jl")
+include("fieldset_broadcast.jl")
 include("interpolate.jl")
 include("ugrid.jl")